+++ CHANGELOG.md
@@ -13,6 +13,12 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
    * Connect backend service /submitservice api added
    * Common component to handle Submit service
    * Show loading icon on click of Save/Save & Proceed button
+* Service Overview
+   * Connect backend service /provided api added
+   * Filter services based on the Status id connected to the /provided api
+   * Search service based on the Offer name connected to the /provided api
+   * Pagination support
+   * Navigate to the Service Details page / Service release process based on the status
 * Service Details Page
 * BugFix:
    * AppReleaseProcess
@@ -21,6 +27,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
       * use case data display fix
    * Service Marketplace
       * Page break issue fix
+   
 
 ## 1.2.0